+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - `fi.evident.dalesbred.SqlQuery` -> `org.dalesbred.query.SqlQuery` 
     - `fi.evident.dalesbred.SqlArray` -> `org.dalesbred.datatype.SqlArray`
     - `fi.evident.dalesbred.ResultTable` -> `org.dalesbred.result.ResultTable`
+    - `fi.evident.support.proxy.TransactionalProxyFactory` -> `org.dalesbred.transaction.TransactionalProxyFactory`
   - Moved classes that are technically `public`, but are not part of Dalesbred's supported 
     API to `org.dalesbred.internal`.
   - Removed `SqlQuery.confidential`. Some database drivers will print the values passed to